refactor(-): :hammer: move serializers dieta especial
