+++ sme_sigpae_api/dieta_especial/solicitacao_dieta_especial/api/viewsets.py
@@ -42,9 +42,30 @@ from sme_sigpae_api.dados_comuns.permissions import (
 )
 from sme_sigpae_api.dados_comuns.services import enviar_email_codae_atualiza_protocolo
 from sme_sigpae_api.dados_comuns.utils import convert_dict_to_querydict
-from sme_sigpae_api.dieta_especial.api.serializers import (
-    AlergiaIntoleranciaSerializer,
+from sme_sigpae_api.dieta_especial.forms import (
+    NegaDietaEspecialForm,
+    RelatorioDietaForm,
+    RelatorioQuantitativoSolicDietaEspForm,
+    SolicitacoesAtivasInativasPorAlunoForm,
+)
+from sme_sigpae_api.dieta_especial.gera_historico_protocolo import (
+    atualiza_historico_protocolo,
+)
+from sme_sigpae_api.dieta_especial.protocolo_padrao.api.serializers import (
     AlimentoSerializer,
+)
+from sme_sigpae_api.dieta_especial.protocolo_padrao.models import (
+    Alimento,
+    ProtocoloPadraoDietaEspecial,
+    SubstituicaoAlimento,
+)
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.api.filters import (
+    AlimentoFilter,
+    DietaEspecialFilter,
+    MotivoNegacaoFilter,
+)
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.api.serializers import (
+    AlergiaIntoleranciaSerializer,
     ClassificacaoDietaSerializer,
     MotivoAlteracaoUESerializer,
     MotivoNegacaoSerializer,
@@ -58,28 +79,9 @@ from sme_sigpae_api.dieta_especial.api.serializers import (
     SolicitacoesAtivasInativasPorAlunoSerializer,
     UnidadeEducacionalSerializer,
 )
-from sme_sigpae_api.dieta_especial.api.serializers_create import (
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.api.serializers_create import (
     AlteracaoUESerializer,
     SolicitacaoDietaEspecialCreateSerializer,
-)
-from sme_sigpae_api.dieta_especial.forms import (
-    NegaDietaEspecialForm,
-    RelatorioDietaForm,
-    RelatorioQuantitativoSolicDietaEspForm,
-    SolicitacoesAtivasInativasPorAlunoForm,
-)
-from sme_sigpae_api.dieta_especial.gera_historico_protocolo import (
-    atualiza_historico_protocolo,
-)
-from sme_sigpae_api.dieta_especial.protocolo_padrao.models import (
-    Alimento,
-    ProtocoloPadraoDietaEspecial,
-    SubstituicaoAlimento,
-)
-from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.api.filters import (
-    AlimentoFilter,
-    DietaEspecialFilter,
-    MotivoNegacaoFilter,
 )
 from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     AlergiaIntolerancia,
