+++ sme_sigpae_api/produto/api/serializers/serializers.py
@@ -3,6 +3,8 @@ import datetime
 import environ
 from rest_framework import serializers
 
+from sme_sigpae_api.dieta_especial.protocolo_padrao.api import serializers as des
+
 from ....dados_comuns.api.serializers import (
     AnexoLogSolicitacoesUsuarioSerializer,
     ContatoSerializer,
@@ -13,7 +15,6 @@ from ....dados_comuns.api.serializers import (
 )
 from ....dados_comuns.fluxo_status import ReclamacaoProdutoWorkflow
 from ....dados_comuns.validators import objeto_nao_deve_ter_duplicidade
-from ....dieta_especial.api import serializers as des
 from ....dieta_especial.protocolo_padrao.models import Alimento
 from ....escola.api.serializers import (
     AlunoSimplesSerializer,
